requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

litellm>=1.40.0
openai>=1.14.0
ragas>=0.2.0
datasets>=2.18.0

sqlalchemy>=2.0.29
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
boto3>=1.34.62

fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.4

python-dotenv>=1.0.1

requests>=2.31.0

# Streamlit frontend
streamlit>=1.32.0
